Move GPU debug step/pause/messages out of Windows/.
