+++ Windows/GEDebugger/GEDebugger.cpp
@@ -19,7 +19,8 @@
 #include <string>
 #include <set>
 
-#include "native/base/mutex.h"
+#include "base/functional.h"
+#include "base/mutex.h"
 #include "Windows/GEDebugger/GEDebugger.h"
 #include "Windows/GEDebugger/SimpleGLWindow.h"
 #include "Windows/GEDebugger/CtrlDisplayListView.h"
@@ -33,40 +34,19 @@
 #include "GPU/Common/GPUDebugInterface.h"
 #include "GPU/GPUState.h"
 #include "GPU/Debugger/Breakpoints.h"
+#include "GPU/Debugger/Stepping.h"
 #include "Core/Config.h"
 #include <windowsx.h>
 #include <commctrl.h>
 
 using namespace GPUBreakpoints;
-
-enum PauseAction {
-	PAUSE_CONTINUE,
-	PAUSE_BREAK,
-	PAUSE_GETFRAMEBUF,
-	PAUSE_GETDEPTHBUF,
-	PAUSE_GETSTENCILBUF,
-	PAUSE_GETTEX,
-	PAUSE_SETCMDVALUE,
-};
+using namespace GPUStepping;
 
 static bool attached = false;
 // TODO
 static bool textureCaching = true;
-static recursive_mutex pauseLock;
-static condition_variable pauseWait;
-static PauseAction pauseAction = PAUSE_CONTINUE;
-static recursive_mutex actionLock;
-static condition_variable actionWait;
-static bool actionComplete;
 
 static BreakNextType breakNext = BREAK_NONE;
-
-static bool bufferResult;
-static GPUDebugBuffer bufferFrame;
-static GPUDebugBuffer bufferDepth;
-static GPUDebugBuffer bufferStencil;
-static GPUDebugBuffer bufferTex;
-static u32 pauseSetCmdValue;
 
 enum PrimaryDisplayType {
 	PRIMARY_FRAMEBUF,
@@ -74,72 +54,9 @@ enum PrimaryDisplayType {
 	PRIMARY_STENCILBUF,
 };
 
-// TODO: Simplify and move out of windows stuff, just block in a common way for everyone.
-
 void CGEDebugger::Init() {
 	SimpleGLWindow::RegisterClass();
 	CtrlDisplayListView::registerClass();
-}
-
-static void SetPauseAction(PauseAction act, bool waitComplete = true) {
-	{
-		lock_guard guard(pauseLock);
-		actionLock.lock();
-		pauseAction = act;
-	}
-
-	actionComplete = false;
-	pauseWait.notify_one();
-	while (waitComplete && !actionComplete) {
-		actionWait.wait(actionLock);
-	}
-	actionLock.unlock();
-}
-
-static void RunPauseAction() {
-	lock_guard guard(actionLock);
-
-	switch (pauseAction) {
-	case PAUSE_CONTINUE:
-		// Don't notify, just go back, woke up by accident.
-		return;
-
-	case PAUSE_BREAK:
-		break;
-
-	case PAUSE_GETFRAMEBUF:
-		bufferResult = gpuDebug->GetCurrentFramebuffer(bufferFrame);
-		break;
-
-	case PAUSE_GETDEPTHBUF:
-		bufferResult = gpuDebug->GetCurrentDepthbuffer(bufferDepth);
-		break;
-
-	case PAUSE_GETSTENCILBUF:
-		bufferResult = gpuDebug->GetCurrentStencilbuffer(bufferStencil);
-		break;
-
-	case PAUSE_GETTEX:
-		bufferResult = gpuDebug->GetCurrentTexture(bufferTex);
-		break;
-
-	case PAUSE_SETCMDVALUE:
-		gpuDebug->SetCmdValue(pauseSetCmdValue);
-		break;
-
-	default:
-		ERROR_LOG(HLE, "Unsupported pause action, forgot to add it to the switch.");
-	}
-
-	actionComplete = true;
-	actionWait.notify_one();
-	pauseAction = PAUSE_BREAK;
-}
-
-static void ForceUnpause() {
-	SetPauseAction(PAUSE_CONTINUE, false);
-	actionComplete = true;
-	actionWait.notify_one();
 }
 
 CGEDebugger::CGEDebugger(HINSTANCE _hInstance, HWND _hParent)
@@ -227,9 +144,9 @@ void CGEDebugger::UpdatePreviews() {
 	// TODO: Do something different if not paused?
 
 	wchar_t desc[256];
-	GPUDebugBuffer *primaryBuffer = NULL;
+	const GPUDebugBuffer *primaryBuffer = NULL;
+	bool bufferResult = false;
 	GPUgstate state;
-	bufferResult = false;
 
 	if (gpuDebug != NULL) {
 		state = gpuDebug->GetGState();
@@ -237,21 +154,24 @@ void CGEDebugger::UpdatePreviews() {
 
 	switch (PrimaryDisplayType(fbTabs->CurrentTabIndex())) {
 	case PRIMARY_FRAMEBUF:
-		SetPauseAction(PAUSE_GETFRAMEBUF);
-		primaryBuffer = &bufferFrame;
-		_snwprintf(desc, ARRAY_SIZE(desc), L"Color: 0x%08x (%dx%d)", state.getFrameBufRawAddress(), primaryBuffer->GetStride(), primaryBuffer->GetHeight());
+		bufferResult = GPU_GetCurrentFramebuffer(primaryBuffer);
+		if (bufferResult) {
+			_snwprintf(desc, ARRAY_SIZE(desc), L"Color: 0x%08x (%dx%d)", state.getFrameBufRawAddress(), primaryBuffer->GetStride(), primaryBuffer->GetHeight());
+		}
 		break;
 
 	case PRIMARY_DEPTHBUF:
-		SetPauseAction(PAUSE_GETDEPTHBUF);
-		primaryBuffer = &bufferDepth;
-		_snwprintf(desc, ARRAY_SIZE(desc), L"Depth: 0x%08x (%dx%d)", state.getDepthBufRawAddress(), primaryBuffer->GetStride(), primaryBuffer->GetHeight());
+		bufferResult = GPU_GetCurrentDepthbuffer(primaryBuffer);
+		if (bufferResult) {
+			_snwprintf(desc, ARRAY_SIZE(desc), L"Depth: 0x%08x (%dx%d)", state.getDepthBufRawAddress(), primaryBuffer->GetStride(), primaryBuffer->GetHeight());
+		}
 		break;
 
 	case PRIMARY_STENCILBUF:
-		SetPauseAction(PAUSE_GETSTENCILBUF);
-		primaryBuffer = &bufferStencil;
-		_snwprintf(desc, ARRAY_SIZE(desc), L"Stencil: 0x%08x (%dx%d)", state.getFrameBufRawAddress(), primaryBuffer->GetStride(), primaryBuffer->GetHeight());
+		bufferResult = GPU_GetCurrentStencilbuffer(primaryBuffer);
+		if (bufferResult) {
+			_snwprintf(desc, ARRAY_SIZE(desc), L"Stencil: 0x%08x (%dx%d)", state.getFrameBufRawAddress(), primaryBuffer->GetStride(), primaryBuffer->GetHeight());
+		}
 		break;
 	}
 
@@ -264,12 +184,12 @@ void CGEDebugger::UpdatePreviews() {
 		SetDlgItemText(m_hDlg, IDC_GEDBG_FRAMEBUFADDR, L"Failed");
 	}
 
-	bufferResult = false;
-	SetPauseAction(PAUSE_GETTEX);
+	const GPUDebugBuffer *bufferTex = NULL;
+	bufferResult = GPU_GetCurrentTexture(bufferTex);
 
 	if (bufferResult) {
-		auto fmt = SimpleGLWindow::Format(bufferTex.GetFormat());
-		texWindow->Draw(bufferTex.GetData(), bufferTex.GetStride(), bufferTex.GetHeight(), bufferTex.GetFlipped(), fmt);
+		auto fmt = SimpleGLWindow::Format(bufferTex->GetFormat());
+		texWindow->Draw(bufferTex->GetData(), bufferTex->GetStride(), bufferTex->GetHeight(), bufferTex->GetFlipped(), fmt);
 
 		if (gpuDebug != NULL) {
 			if (state.isTextureAlphaUsed()) {
@@ -336,7 +256,7 @@ void CGEDebugger::SetBreakNext(BreakNextType type) {
 	attached = true;
 	SetupPreviews();
 
-	SetPauseAction(PAUSE_CONTINUE, false);
+	ResumeFromStepping();
 	breakNext = type;
 }
 
@@ -436,7 +356,7 @@ BOOL CGEDebugger::DlgProc(UINT message, WPARAM wParam, LPARAM lParam) {
 			SetDlgItemText(m_hDlg, IDC_GEDBG_TEXADDR, L"");
 
 			// TODO: detach?  Should probably have separate UI, or just on activate?
-			SetPauseAction(PAUSE_CONTINUE, false);
+			ResumeFromStepping();
 			breakNext = BREAK_NONE;
 			break;
 		}
@@ -486,8 +406,7 @@ BOOL CGEDebugger::DlgProc(UINT message, WPARAM wParam, LPARAM lParam) {
 
 	case WM_GEDBG_SETCMDWPARAM:
 		{
-			pauseSetCmdValue = (u32)wParam;
-			SetPauseAction(PAUSE_SETCMDVALUE);
+			GPU_SetCmdValue((u32)wParam);
 		}
 		break;
 	}
@@ -501,23 +420,12 @@ bool WindowsHost::GPUDebuggingActive() {
 	return attached;
 }
 
-static void PauseWithMessage(UINT msg, WPARAM wParam = NULL, LPARAM lParam = NULL) {
-	lock_guard guard(pauseLock);
-	if (coreState != CORE_RUNNING && coreState != CORE_NEXTFRAME) {
-		return;
-	}
-
+static void DeliverMessage(UINT msg, WPARAM wParam, LPARAM lParam) {
 	PostMessage(geDebuggerWindow->GetDlgHandle(), msg, wParam, lParam);
+}
 
-	// Just to be sure.
-	if (pauseAction == PAUSE_CONTINUE) {
-		pauseAction = PAUSE_BREAK;
-	}
-
-	do {
-		RunPauseAction();
-		pauseWait.wait(pauseLock);
-	} while (pauseAction != PAUSE_CONTINUE);
+static void PauseWithMessage(UINT msg, WPARAM wParam = NULL, LPARAM lParam = NULL) {
+	EnterStepping(std::bind(&DeliverMessage, msg, wParam, lParam));
 }
 
 void WindowsHost::GPUNotifyCommand(u32 pc) {
